### 2016-09-16 #63 @whedon
[REVIEW]: rintrojs
**Submitting author:** @carlganz (<a href="http://orcid.org/0000-0002-2973-3545">Carl Ganz</a>)
**Repository:** <a href="https://github.com/carlganz/rintrojs" target ="_blank">https://github.com/carlganz/rintrojs</a>
**Version:** v0.1.1
**Archive:** <a href="https://dx.doi.org/10.5281/zenodo.160211" target ="_blank">https://dx.doi.org/10.5281/zenodo.160211</a>
