+++ DESCRIPTION
@@ -1,7 +1,7 @@
 Package: rintrojs
 Type: Package
 Title: Wrapper for the 'Intro.js' Library
-Version: 0.1.0
+Version: 0.1.0.900
 Authors@R: c(
   person("Carl", "Ganz", email = "carlganz@gmail.com", role = c("aut", "cre")),
   person("Afshin", "Mehrabani", role = c("ctb","cph"), comment = "intro.js in javascript/introjs")
