### 2016-09-19 carlganz/rintrojs#8 @brylie
License.txt
I am reviewing this package as part of a [submission to the Journal of Open Source Software](https://github.com/openjournals/joss-reviews/issues/63).

As part of my review, I am instructed to locate a license file for the sourcecode. Where can I locate the LICENSE.text, or equivalent file for your wrapper package code?
