### 2016-09-19 #63 @brylie
Could not locate LICENSE.txt, or equivalent (for wrapper project):

- https://github.com/carlganz/rintrojs/issues/8
