### 2016-09-19 carlganz/rintrojs#9 @brylie
Release version for code?
I am reviewing this package as part of a [submission to the Journal of Open Source Software](https://github.com/openjournals/joss-reviews/issues/63).

As part of my review, I am instructed to locate a release package for this wrapper, but the [Github project releases page](https://github.com/carlganz/rintrojs/releases) is empty. Where can I find a release version (0.1.0) for this package?
