### 2016-09-19 #63 @brylie
Could not locate a Github release for 0.1.0, or any release at all:

- https://github.com/carlganz/rintrojs/issues/9
- https://github.com/carlganz/rintrojs/releases
