### 2016-09-19 carlganz/rintrojs#10 @brylie
Improve contributing guidelines
I am reviewing this package as part of a [submission to the Journal of Open Source Software](https://github.com/openjournals/joss-reviews/issues/63).

As part of my review, I am instructed to locate contributing guidelines for this project. The contributing guidelines state "*All feedback is appreciated.*"

Please add specific details as to how contributors can:

1) Contribute to the software
2) Report issues or problems with the software
3) Seek support
