### 2016-09-19 #63 @brylie
Contributing guidelines were a bit 'thin':

- https://github.com/carlganz/rintrojs/issues/10
