### 2016-09-19 #63 @brylie
There are no 'performance claims' made by this software, so leaving 'performance' checkbox unchecked.
