### 2016-09-19 carlganz/rintrojs#11 @brylie
Statement of need
I am reviewing this package as part of a [submission to the Journal of Open Source Software](https://github.com/openjournals/joss-reviews/issues/63). As part of my review, I am instructed to locate the statement of need for this project.

> Do the authors clearly state what problems the software is designed to solve and who the target audience is?

I have reviewed the published PDF as well as README for this project. The PDF alludes to a statement of need in the second paragraph, but could be improved. Please provide a clear statement of need for this project.
