### 2016-09-19 #63 @brylie
The **Statement of Need** for the project could be improved:

- https://github.com/carlganz/rintrojs/issues/11

In effect, the statement of need is provided in the second paragraph of the published PDF. However, the statement of need is brief, and not clearly distinguished from other content.
