### 2016-09-19 #63 @brylie
**DOI References**: the paper contains two citations, neither of which has a DOI number. I am not sure how to determine whether a reference *should* have a DOI number. Assistance needed.
