### 2016-09-19 carlganz/rintrojs#12 @brylie
DOI numbers in paper references
I am reviewing this package as part of a [submission to the Journal of Open Source Software](https://github.com/openjournals/joss-reviews/issues/63). As part of my review, I am instructed to locate DOI number(s) for all relevant references in the publication.

> Do all archival references that should have a DOI list one (e.g. papers, datasets, software)?

Please provide DOI numbers for the following references:

- Chang, Winston, Joe Cheng, JJ Allaire, Yihui Xie, and Jonathan McPherson. 2016. Shiny: Web Application Framework for R. http://CRAN.R-project.org/package=shiny.
- Mehrabani, Afshin. 2016. Intro.js: Better Introductions for Websites and Features with a Step-by-Step Guide for Your Projects. http://introjs.com.
