### 2016-09-19 carlganz/rintrojs#12 @brylie
Consider the free Zenodo service for creating DOI numbers:
https://zenodo.org/
