### 2016-09-19 #63 @danielskatz
If there are no performance claims, you can check the performance claims box, since the answer is more or less yes.

For the papers, you should be able to look them up in google or google scholar or other means, go to the place they are published or archived, and see if they do or do not have DOIs.
