### 2016-09-19 carlganz/rintrojs#8 @arfon
@brylie - R packages usually include this in the `DESCRIPTION` file (see [here](https://github.com/carlganz/rintrojs/blob/7c069348b53a017ea3ef39a42b5f9a0e91bac371/DESCRIPTION#L13)). While this isn't ideal, CRAN (the R package manager) won't permit plain text LICENSE files.
