### 2016-09-19 carlganz/rintrojs#8 @carlganz
I can add a licence file to the github repo if you'd like.

Thanks for the review everyone!
