### 2016-09-19 carlganz/rintrojs#12 @carlganz
I neither of the softwares I reference have DOI numbers.
