### 2016-09-19 carlganz/rintrojs#8 @arfon
> I can add a licence file to the github repo if you'd like.

Sounds good to me @carlganz. Thanks.
