### 2016-09-20 #63 @brylie
> For the papers, you should be able to look them up in google or google scholar or other means, go to the place they are published or archived, and see if they do or do not have DOIs.

The citations were both links to the software websites.
