### 2016-09-20 #63 @danielskatz
Then there are no DOIs for those products, and you can check the review box that this is ok.
