+++ R/introjs.R
@@ -1,6 +1,19 @@
-### rintrojs package
-### written by Carl Ganz
-### wrapper for the introjs javascript library written by Afshin Mehrabani
+# Copyright (C) 2016 Carl Ganz
+#
+# This file is part of rintrojs.
+# 
+# rintrojs is free software: you can redistribute it and/or modify
+# it under the terms of the GNU General Public License as published by
+# the Free Software Foundation, either version 3 of the License, or
+# (at your option) any later version.
+# 
+# rintrojs is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+# GNU General Public License for more details.
+# 
+# You should have received a copy of the GNU General Public License
+# along with rintrojs.  If not, see <http://www.gnu.org/licenses/>.
 
 #' Initiate intro.js
 #'
