### 2016-09-21 carlganz/rintrojs#8 @carlganz
Added license
