### 2016-09-22 carlganz/rintrojs#8 @graingert
FYI this should be the AGPL-3
