+++ NEWS.md
@@ -4,3 +4,4 @@
 
 * Upgrade to Intro.js 2.3.0
 
+* Use AGPL-3 license (Thanks @graingert #8)
