+++ README.md
@@ -4,9 +4,9 @@
 rintrojs
 ========
 
-The goal of `rintrojs` is to integrate the javascript library [`Intro.js`](http://introjs.com/) into `shiny`.
+[R](https://r-project.org) was originally developed with the name S in the 70's as a glue language. Statisticians at Bell Labs needed an interactive environment for working with their C and Fortran scripts. Decades later, R is still a glue language, except instead of C and Fortran, it brings together C++, and JavaScript. With the advent of the [Shiny](http://http://shiny.rstudio.com/) package, R is now a popular platform for developing data-driven web applications. As Shiny increases in popularity so will the complexity of the Shiny apps built. In many instances, new-users will require guidance when they first use a Shiny application.
 
-All credit for `Intro.js` goes to its author Afshin Mehrabani. I have contributed nothing to `Intro.js`. I simply wrote a wrapper for easy integration with `shiny`.
+Luckily, there is already a well established JavaScript library for this purpose. [Intro.js](http://introjs.com), written by Afshin Mehrabani, is a JavaScript library that helps integrate step-by-step introductions, and clickable hints into websites. The `rintrojs` R package integrates Intro.js into Shiny, so that users can easily add instructions to their application without having to use HTML, CSS, or JavaScript.
 
 Install
 -------
