### 2016-09-27 3e1cc09
add travis r tests
