+++ .travis.yml
@@ -5,6 +5,11 @@ sudo: false
 cache: packages
 warnings_are_errors: true
 
+r:
+  - oldrel
+  - release
+  - devel
+
 os:
   - linux
   - osx
