### 2016-09-27 #63 @brylie
A Github release was just made:

https://github.com/carlganz/rintrojs/releases
