### 2016-09-27 #63 @brylie
The project license was updated:

https://github.com/carlganz/rintrojs/commit/ae3137d2168fa8937dae1c2fb708895ce566338e
