### 2016-09-27 #63 @brylie
Contribution guidelines and code of conduct were just added:

https://github.com/carlganz/rintrojs/commit/7d6bf2d662a16325618910506ff669281ce10d98
