### 2016-09-27 #63 @brylie
Statement of Need was added:

https://github.com/carlganz/rintrojs/commit/d3ef95ad448bac7d63f74899fff6269c93484186
