### 2016-09-28 b4addaa
Don't test dev r on mac
