+++ README.md
@@ -129,7 +129,7 @@ ui <- shinyUI(fluidPage(
 
 server <- shinyServer(function(input, output, session) {
    
-  steps <- reactive(data.frame(element = c(NA,"btn"),
+  steps <- reactive(data.frame(element = c(NA,"#btn"),
                                intro = c(input$intro,"This is a button")))
   
   observeEvent(input$btn,{
