### 2016-10-07 e4cc88a
first attempt at shinytest
