+++ tests/testthat/apps/app1/server.R
@@ -1,0 +1,14 @@
+library(shiny)
+library(rintrojs)
+
+shinyServer(function(input, output, session) {
+  
+  steps <- reactive(data.frame(element = c(NA,"#btn"),
+                               intro = c(input$intro,"This is a button")))
+  
+  observeEvent(input$btn,{
+    introjs(session,options = list(steps=steps()))
+    
+  })
+  
+})
