+++ tests/testthat/apps/app1/ui.R
@@ -1,0 +1,11 @@
+library(shiny)
+library(rintrojs)
+
+shinyUI(fluidPage(
+  introjsUI(),
+  mainPanel(
+    textInput("intro","Enter an introduction"),
+    actionButton("btn","Press me")
+  )
+)
+)
