+++ tests/testthat/apps/app1/DESCRIPTION
@@ -1,0 +1,5 @@
+Title: rintrojs Programmatic
+Author: Carl Ganz
+License: MIT
+Tags: rintrojs
+Type: Shiny
