+++ .travis.yml
@@ -5,6 +5,11 @@ sudo: false
 cache: packages
 warnings_are_errors: true
 
+r_github_packages:
+  - gaborcsardi/debugme
+  - MangoTheCat/processx
+  - MangoTheCat/webdriver
+
 matrix:
   include:
     - os: linux
