### 2016-10-07 b9e5550
Don't suggest shinytest
