+++ DESCRIPTION
@@ -18,4 +18,4 @@ RoxygenNote: 5.0.1
 URL: https://github.com/carlganz/rintrojs
 BugReports: https://github.com/carlganz/rintrojs/issues
 Suggests: testthat,
-    covr, shinytest
+    covr
