### 2016-10-07 07b07f3
actually install shinytest on travis
