+++ .travis.yml
@@ -9,6 +9,7 @@ r_github_packages:
   - gaborcsardi/debugme
   - MangoTheCat/processx
   - MangoTheCat/webdriver
+  - MangoTheCat/shinytest
 
 matrix:
   include:
