+++ tests/testthat.R
@@ -1,5 +1,4 @@
 library(testthat)
-library(shinytest)
 library(rintrojs)
 
 test_check("rintrojs")
