### 2016-10-10 #63 @brylie
@arfon I thought I was no longer assigned to the review, since my name was removed from the 'Assignees' field. Also, I can no longer edit the checkboxes in the task description.

What should I do to continue/finish this review?
