### 2016-10-11 carlganz/rintrojs#14 @brylie
Please make release with review changes
Now that we are nearly done with the review process, it would be good to make a release incorporating all recent changes. Please make one more release, so that we may finish the review.
