### 2016-10-11 #63 @brylie
@arfon the review is complete, and all boxes are checked. I requested that the upstream author make one more release, to include changes from the review process:

https://github.com/carlganz/rintrojs/issues/14
