### 2016-10-11 c9bc456
remove shinytests for now
