+++ NEWS.md
@@ -1,7 +1,9 @@
-# rintrojs 0.1.0.900
+# rintrojs 0.1.1
 
 * Fix events bug (#7)
 
 * Upgrade to Intro.js 2.3.0
 
 * Use AGPL-3 license (Thanks @graingert #8)
+
+* Meet Journal of Open Source Software requirements
