### 2016-10-11 carlganz/rintrojs#14 @carlganz
https://github.com/carlganz/rintrojs/releases/tag/0.1.1

Thanks for your help.
