### 2016-10-11 #63 @arfon
> @arfon the review is complete, and all boxes are checked. I requested that the upstream author make one more release, to include changes from the review process:

👍 excellent thanks.

@carlganz. Once you've addressed https://github.com/carlganz/rintrojs/issues/14 could you make an archive of the reviewed software in Zenodo/figshare/other service and update this thread with the DOI of the archive? I can then move forward with accepting the submission.
