### 2016-10-11 #63 @carlganz
Thanks @brylie and @arfon!

Here is the release that should close carlganz/rintrojs#14: https://github.com/carlganz/rintrojs/releases/tag/0.1.1

And here is the archive on Zenodo: https://zenodo.org/record/160211#.V_1As4QrLcs
