### 2016-10-12 #63 @arfon
Thanks for the review @brylie.

@carlganz - your paper is now accepted into JOSS. Your DOI is http://dx.doi.org/10.21105/joss.00063 🚀 🎉 💥
